+++ components/dashboard-table.tsx
@@ -40,6 +40,7 @@ import {
   TooltipProvider,
   TooltipTrigger,
 } from "@/components/ui/tooltip";
+import { DataTablePagination } from "./datatables/pagination";
 
 export type CommandTrigger = {
   serverTrigger: GameServerTrigger;
@@ -266,23 +267,7 @@ export default function DashboardTable<TData, TValue>({
           </TableBody>
         </Table>
         <div className="flex items-center justify-end space-x-2 py-4">
-          <Button
-            variant="outline"
-            size="sm"
-            onClick={() => table.previousPage()}
-            disabled={!table.getCanPreviousPage()}
-          >
-            Previous
-          </Button>
-          <Button
-            variant="outline"
-            size="sm"
-            onClick={() => table.nextPage()}
-            disabled={!table.getCanNextPage()}
-          >
-            Next
-          </Button>
-        </div>
+          <DataTablePagination table={table}/>        </div>
       </div>
     </div>
   );